+++ src/components/alert/alert.component.ts
@@ -1,4 +1,5 @@
 import { animateTo, stopAnimations } from '../../internal/animate.js';
+import { blurActiveElement } from '../../internal/closeActiveElement.js';
 import { classMap } from 'lit/directives/class-map.js';
 import { getAnimation, setDefaultAnimation } from '../../utilities/animation-registry.js';
 import { HasSlotController } from '../../internal/slot.js';
@@ -157,6 +158,7 @@ export default class SlAlert extends ShoelaceElement {
       this.emit('sl-after-show');
     } else {
       // Hide
+      blurActiveElement(this);
       this.emit('sl-hide');
 
       clearTimeout(this.autoHideTimeout);