+++ src/components/alert/alert.test.ts
@@ -123,20 +123,97 @@ describe('<sl-alert>', () => {
   });
 
   describe('close button', () => {
-    it('shows a close button if the alert has the closable attribute', () => async () => {
+    it('shows a close button if the alert has the closable attribute', async () => {
       const alert = await fixture<SlAlert>(html` <sl-alert open closable>I am an alert</sl-alert> `);
       const closeButton = getCloseButton(alert);
 
       expect(closeButton).to.be.visible;
     });
 
-    it('clicking the close button closes the alert', () => async () => {
+    it('clicking the close button closes the alert', async () => {
       const alert = await fixture<SlAlert>(html` <sl-alert open closable>I am an alert</sl-alert> `);
       const closeButton = getCloseButton(alert);
 
       await expectHideAndAfterHideToBeEmittedInCorrectOrder(alert, () => {
         clickOnElement(closeButton!);
       });
+    });
+
+    it('clicking above close button does not close the alert', async () => {
+      const wrapper = await fixture<HTMLDivElement>(
+        html`<div class="wrapper" style="padding: 24px; background-color:red;">
+          <sl-alert open closable>I am an alert</sl-alert>
+        </div>`
+      );
+      const alert = wrapper.querySelector('sl-alert')!;
+
+      const clickTargetPromise = new Promise<HTMLElement>(resolve => {
+        const clickHandler = sinon.spy((event: MouseEvent) => {
+          resolve(event.target as HTMLElement);
+        });
+        alert.shadowRoot!.addEventListener('click', clickHandler);
+        wrapper.addEventListener('click', clickHandler);
+      });
+
+      const closeButton = getCloseButton(alert);
+      await clickOnElement(closeButton!, 'top', 0, -4);
+      const clickTarget = await clickTargetPromise;
+      await expect(clickTarget.tagName.toLowerCase()).to.not.be.equal('sl-icon-button');
+      expect(clickTarget.classList.contains('alert')).to.be.true;
+      expect(clickTarget.classList.contains('wrapper'), 'The click should happen in the alert and not outside of it').to
+        .be.false;
+    });
+
+    it('clicking under close button does not close the alert', async () => {
+      const wrapper = await fixture<HTMLDivElement>(
+        html`<div class="wrapper" style="padding: 24px; background-color:red;">
+          <sl-alert open closable>I am an alert</sl-alert>
+        </div>`
+      );
+      const alert = wrapper.querySelector('sl-alert')!;
+
+      const clickTargetPromise = new Promise<HTMLElement>(resolve => {
+        const clickHandler = sinon.spy((event: MouseEvent) => {
+          resolve(event.target as HTMLElement);
+        });
+        alert.shadowRoot!.addEventListener('click', clickHandler);
+        wrapper.addEventListener('click', clickHandler);
+      });
+
+      const closeButton = getCloseButton(alert);
+      await clickOnElement(closeButton!, 'bottom', 0, 4);
+      const clickTarget = await clickTargetPromise;
+
+      await expect(clickTarget.tagName.toLowerCase()).to.not.be.equal('sl-icon-button');
+      expect(clickTarget.classList.contains('alert')).to.be.true;
+      expect(clickTarget.classList.contains('wrapper'), 'The click should happen in the alert and not outside of it').to
+        .be.false;
+    });
+
+    it('clicking on the right side of the close button does not close the alert', async () => {
+      const wrapper = await fixture<HTMLDivElement>(
+        html`<div class="wrapper" style="padding: 24px; background-color:red;">
+          <sl-alert open closable>I am an alert</sl-alert>
+        </div>`
+      );
+      const alert = wrapper.querySelector('sl-alert')!;
+
+      const clickTargetPromise = new Promise<HTMLElement>(resolve => {
+        const clickHandler = sinon.spy((event: MouseEvent) => {
+          resolve(event.target as HTMLElement);
+        });
+        alert.shadowRoot!.addEventListener('click', clickHandler);
+        wrapper.addEventListener('click', clickHandler);
+      });
+
+      const closeButton = getCloseButton(alert);
+      await clickOnElement(closeButton!, 'right', 4, 0);
+      const clickTarget = await clickTargetPromise;
+
+      await expect(clickTarget.tagName.toLowerCase()).to.not.be.equal('sl-icon-button');
+      expect(clickTarget.classList.contains('alert')).to.be.true;
+      expect(clickTarget.classList.contains('wrapper'), 'The click should happen in the alert and not outside of it').to
+        .be.false;
     });
   });