+++ docs/pages/resources/changelog.md
@@ -12,6 +12,15 @@ Components with the <sl-badge variant="warning" pill>Experimental</sl-badge> bad
 
 New versions of Shoelace are released as-needed and generally occur when a critical mass of changes have accumulated. At any time, you can see what's coming in the next release by visiting [next.shoelace.style](https://next.shoelace.style).
 
+## 2.20.1
+
+- Fixed a bug that prevented `<sl-tab-group>` to be activated properly when rendered in another `<sl-tab-group>` [#2367]
+- Fixed a bug that in `<sl-dropdown>` that prevented tab from working properly in some cases [#2371]
+- Fixed the guard on popover to allow virtual elements [#2399]
+- Fixed the close button in `<sl-alert>` so clicking above/below it doesn't inadvertently close it [#2375]
+- Fixed accessibility issues for elements that are closed while having slotted focused children. [#2383]
+- Improved accessibility of `<sl-carousel>` [#2364]
+
 ## 2.20.0
 
 - Added the ability to set a custom snap function and use `repeat(n)` to `<sl-split-panel>` [#2340]